move ChildNeedsUpdating() function since it's not just related to Progressive logic
