+++ internal/controller/monovertexrollout/monovertexrollout_progressive.go
@@ -13,8 +13,6 @@ import (
 
 	"github.com/numaproj/numaplane/internal/common"
 	ctlrcommon "github.com/numaproj/numaplane/internal/controller/common"
-	"github.com/numaproj/numaplane/internal/controller/common/numaflowtypes"
-	"github.com/numaproj/numaplane/internal/util"
 )
 
 // Implemented functions for the progressiveController interface:
@@ -80,31 +78,6 @@ func (r *MonoVertexRolloutReconciler) Recycle(ctx context.Context,
 		return false, err
 	}
 	return true, nil
-}
-
-// ChildNeedsUpdating() tests for essential equality, with any irrelevant fields eliminated from the comparison
-func (r *MonoVertexRolloutReconciler) ChildNeedsUpdating(ctx context.Context, from, to *unstructured.Unstructured) (bool, error) {
-	numaLogger := logger.FromContext(ctx)
-	// remove "replicas" field from comparison to test for equality
-	fromNew, err := numaflowtypes.MonoVertexWithoutReplicas(from)
-	if err != nil {
-		return false, err
-	}
-	toNew, err := numaflowtypes.MonoVertexWithoutReplicas(to)
-	if err != nil {
-		return false, err
-	}
-
-	specsEqual := util.CompareStructNumTypeAgnostic(fromNew, toNew)
-	numaLogger.Debugf("specsEqual: %t, fromNew=%v, toNew=%v\n",
-		specsEqual, fromNew, toNew)
-	labelsEqual := util.CompareMaps(from.GetLabels(), to.GetLabels())
-	numaLogger.Debugf("labelsEqual: %t, from Labels=%v, to Labels=%v", labelsEqual, from.GetLabels(), to.GetLabels())
-	annotationsEqual := util.CompareMaps(from.GetAnnotations(), to.GetAnnotations())
-	numaLogger.Debugf("annotationsEqual: %t, from Annotations=%v, to Annotations=%v", annotationsEqual, from.GetAnnotations(), to.GetAnnotations())
-
-	return !specsEqual || !labelsEqual || !annotationsEqual, nil
-
 }
 
 // AssessUpgradingChild makes an assessment of the upgrading child to determine if it was successful, failed, or still not known
